+++ src/pages/SubmissionConfig.js
@@ -63,7 +63,6 @@ export const SubmissionConfig = ({ methodId }) => {
   // TODO: this should probably be moved to a scope more local to the data selector
   const [recordsTableSort, setRecordsTableSort] = useState({ field: 'id', direction: 'asc' })
   const [inputTableSort, setInputTableSort] = useState({ field: '', direction: 'asc' })
-  const [outputTableSort, setOutputTableSort] = useState({ field: '', direction: 'asc' })
 
   const [displayLaunchModal, setDisplayLaunchModal] = useState(false)
   const [noRecordTypeData, setNoRecordTypeData] = useState(null)
@@ -419,8 +418,7 @@ export const SubmissionConfig = ({ methodId }) => {
   const renderOutputs = () => {
     return configuredOutputDefinition ? h(OutputsTable, {
       selectedDataTable: _.keyBy('name', recordTypes)[selectedRecordType],
-      configuredOutputDefinition, setConfiguredOutputDefinition,
-      outputTableSort, setOutputTableSort
+      configuredOutputDefinition, setConfiguredOutputDefinition
     }) : 'No previous run set data...'
   }
