+++ src/pages/SubmissionConfig.test.js
@@ -1566,79 +1566,6 @@ describe('SubmissionConfig inputs/outputs definitions', () => {
     within(cells2[3]).getByDisplayValue('target_workflow_1_file_output')
   })
 
-  it('should set output variable names when set defaults button is clicked', async () => {
-    const mockRunSetResponse = jest.fn(() => Promise.resolve(runSetResponse))
-    const mockMethodsResponse = jest.fn(() => Promise.resolve(methodsResponse))
-    const mockSearchResponse = jest.fn((_, recordType) => Promise.resolve(searchResponses[recordType]))
-    const mockTypesResponse = jest.fn(() => Promise.resolve(typesResponse))
-
-    await Ajax.mockImplementation(() => {
-      return {
-        Cbas: {
-          runSets: {
-            getForMethod: mockRunSetResponse
-          },
-          methods: {
-            getById: mockMethodsResponse
-          }
-        },
-        Wds: {
-          search: {
-            post: mockSearchResponse
-          },
-          types: {
-            get: mockTypesResponse
-          }
-        }
-      }
-    })
-
-    render(h(SubmissionConfig))
-
-    await waitFor(() => {
-      expect(mockRunSetResponse).toHaveBeenCalledTimes(1)
-      expect(mockTypesResponse).toHaveBeenCalledTimes(1)
-      expect(mockMethodsResponse).toHaveBeenCalledTimes(1)
-      expect(mockSearchResponse).toHaveBeenCalledTimes(1)
-    })
-
-    const button = await screen.findByRole('button', { name: 'Outputs' })
-    await fireEvent.click(button)
-
-    const table = await screen.findByRole('table')
-    const rows = within(table).queryAllByRole('row')
-    const headers = within(rows[0]).queryAllByRole('columnheader')
-    const cells1 = within(rows[1]).queryAllByRole('cell')
-    const cells2 = within(rows[2]).queryAllByRole('cell')
-
-    // prepopulated values
-    within(cells1[0]).getByText('target_workflow_1')
-    within(cells1[1]).getByText('file_output')
-    within(cells1[2]).getByText('File')
-    within(cells1[3]).getByDisplayValue('target_workflow_1_file_output')
-
-    within(cells2[0]).getByText('target_workflow_1')
-    within(cells2[1]).getByText('unused_output')
-    within(cells2[2]).getByText('String')
-    within(cells2[3]).getByDisplayValue('')
-
-    // set defaults
-    await act(async () => {
-      await fireEvent.click(within(headers[3]).getByRole('button'))
-    })
-
-    // default values set
-    within(cells1[0]).getByText('target_workflow_1')
-    within(cells1[1]).getByText('file_output')
-    within(cells1[2]).getByText('File')
-    within(cells1[3]).getByDisplayValue('file_output')
-
-    within(cells2[0]).getByText('target_workflow_1')
-    within(cells2[1]).getByText('unused_output')
-    within(cells2[2]).getByText('String')
-    within(cells2[3]).getByDisplayValue('unused_output')
-  })
-
   it('should display struct builder modal when "view struct builder" link is clicked', async () => {
     // ** ARRANGE **
     const mockRunSetResponse = jest.fn(() => Promise.resolve(runSetResponseWithStruct))
